chore: remove stepItemRender component
