+++ renderer/src/pages/Dashboard/index.module.scss
@@ -13,3 +13,28 @@
     }
   }
 }
+
+.customNode {
+  height: 100%;
+  width: 100%;
+  border-radius: 50%;
+  border: 1px solid #3e71f1;
+  text-align: center;
+  background: #3e71f1;
+  color: #fff;
+  position: relative;
+
+  i, span {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    left: 0;
+    width: 100%;
+    text-align: center;
+  }
+}
+
+.activeNode {
+  background: #fff;
+  color: #3e71f1;
+}
